Move autofocus manager to its own file
